Add organization to server repo
