+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "redis-server | ts-node-dev app/app.ts",
+    "start": "redis-server | ts-node-dev app.ts",
     "build": "docker-compose up",
     "tsc": "tsc"
   },
