+++ src/redis/redis-client.ts
@@ -1,4 +1,6 @@
 // redis-client.js
+require("dotenv").config()
+
 const redis = require("redis")
 const { promisify } = require("util")
 const client = redis.createClient(process.env.REDIS_URL)
